app/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
bitarray==3.0.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cytoolz==1.0.1
Django==5.1.3
django-baton==4.1.0
django-cors-headers==4.6.0
django-ninja==1.3.0
django-adminlte3==0.1.6
django-jazzmin==2.6.0
eth-account==0.13.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.1.0
eth_abi==5.2.0
frozendict==2.3.10
frozenlist==1.5.0
hexbytes==1.3.0
idna==3.10
kombu==5.4.2
moralis==0.1.49
multidict==6.1.0
parsimonious==0.10.0
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2==2.9.10
pycryptodome==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
pyunormalize==16.0.0
qrcode==8.0
redis==5.2.0
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
sentry-sdk==2.20.0
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uWSGI==2.0.28
vine==5.1.0
wcwidth==0.2.13
web3==7.7.0
websockets==13.1
yarl==1.18.3
